+++ src/webapp/src/components/pipelines/PipelineSchedules.vue
@@ -43,7 +43,13 @@ export default {
   },
   methods: {
     ...mapActions('orchestration', ['deletePipelineSchedule']),
-    deletePipeline(pipeline) {
+    goToCreatePipeline() {
+      this.$router.push({ name: 'createPipelineSchedule' })
+    },
+    goToLog(jobId) {
+      this.$router.push({ name: 'runLog', params: { jobId } })
+    },
+    removePipeline(pipeline) {
       this.deletePipelineSchedule(pipeline)
         .then(() =>
           Vue.toasted.global.success(
@@ -51,12 +57,6 @@ export default {
           )
         )
         .catch(error => Vue.toasted.global.error(error.response.data.code))
-    },
-    goToCreatePipeline() {
-      this.$router.push({ name: 'createSchedule' })
-    },
-    goToLog(jobId) {
-      this.$router.push({ name: 'runLog', params: { jobId } })
     },
     runELT(pipeline) {
       this.$store.dispatch('orchestration/run', pipeline)
@@ -90,15 +90,12 @@ export default {
 
     <div class="columns is-vcentered">
       <div class="column">
-        <h2 class="title is-5">Pipelines</h2>
-      </div>
-
-      <div class="column">
-        <div class="field is-pulled-right">
+        <h2 class="title is-inline-block">Pipelines</h2>
+        <div class="field is-pulled-right is-inline-block">
           <div class="control">
             <button
-              class="button is-interactive-primary"
-              @click="goToCreatePipeline()"
+              class="button is-medium is-interactive-primary"
+              @click="goToCreatePipeline"
             >
               <span>Create</span>
             </button>
@@ -232,7 +229,7 @@ export default {
                           <button
                             class="button is-danger"
                             data-dropdown-auto-close
-                            @click="deletePipeline(pipeline)"
+                            @click="removePipeline(pipeline)"
                           >
                             Delete
                           </button>