+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 ### New
 
 - [#1461](https://gitlab.com/meltano/meltano/issues/1461) Display toasted notification for report adding to dashboard
+- [#1419](https://gitlab.com/meltano/meltano/issues/1419) Add ability to edit and delete dashboards
 
 ### Changes